+++ packages/svelte-ux/src/lib/styles/theme.ts
@@ -1,5 +1,16 @@
 import { range } from 'd3-array';
 
+import {
+  rgb,
+  hsl,
+  oklch,
+  clampRgb,
+  interpolate,
+  wcagContrast,
+  formatCss,
+  type Color,
+} from 'culori';
+
 export const semanticColors = ['primary', 'secondary', 'accent', 'neutral'] as const;
 export const stateColors = ['info', 'success', 'warning', 'danger'] as const;
 export const colors = [...semanticColors, ...stateColors];
@@ -20,6 +31,170 @@ export const colorNames = [
   'surface-300',
   'surface-content',
 ];
+
+/**
+ * Generate theme colors (ex. { primary: hsl(var(--color-primary) / <alpha-value>), ... })
+ */
+export function createThemeColors(colorSpace: 'rgb' | 'hsl' | 'oklch') {
+  return Object.fromEntries(
+    colorNames.map((color) => [color, `${colorSpace}(var(--color-${color}) / <alpha-value>)`])
+  );
+}
+
+/**
+ * Convert names to CSS variables and color values common color space (hsl, oklch, etc) and space separated
+ */
+export function processThemeColors(
+  themeColors: Record<string, string>,
+  colorSpace: 'rgb' | 'hsl' | 'oklch'
+) {
+  const colors = { ...themeColors };
+
+  // Generate optional semanatic colors
+  if (!('neutral' in colors) && !('neutral-500' in colors)) {
+    colors['neutral'] = 'oklch(.355192 .032071 262.988584)';
+  }
+
+  // Generate optional state colors
+  if (!('info' in colors) && !('info-500' in colors)) {
+    colors['info'] = 'oklch(0.7206 0.191 231.6)';
+  }
+  if (!('success' in colors) && !('success-500' in colors)) {
+    colors['success'] = 'oklch(64.8% 0.150 160)';
+  }
+  if (!('warning' in colors) && !('warning-500' in colors)) {
+    colors['warning'] = 'oklch(0.8471 0.199 83.87)';
+  }
+  if (!('danger' in colors) && !('danger-500' in colors)) {
+    colors['danger'] = 'oklch(0.7176 0.221 22.18)';
+  }
+
+  // Generate optional content colors
+  for (const color of [...semanticColors, ...stateColors]) {
+    // Add `primary` from `primary-500` if not defined in theme (ex. Skeleton)
+    if (!(color in colors) && `${color}-500` in themeColors) {
+      colors[color] = themeColors[`${color}-500`];
+    }
+
+    if (!(`${color}-content` in colors)) {
+      colors[`${color}-content`] = foregroundColor(colors[color]);
+    }
+
+    // Generate color shades (ex. `primary-500`) if not defined.  Useful for Daisy but not Skeleton themes, for example
+    for (const shade of shades) {
+      const shadeColorName = `${color}-${shade}`;
+      if (!(shadeColorName in colors)) {
+        const newColor =
+          shade < 500
+            ? lightenColor(colors[color], (500 - shade) / 1000) // 100 == 0.1
+            : shade > 500
+              ? darkenColor(colors[color], (shade - 500) / 1000) // 100 == 0.1
+              : colors[color];
+        colors[shadeColorName] = newColor;
+      }
+    }
+  }
+
+  // Generate optional surface colors
+  if (!('surface-100' in colors)) {
+    colors['surface-100'] = 'oklch(100 0 0)';
+  }
+
+  if (!('surface-200' in colors)) {
+    colors['surface-200'] = darkenColor(colors['surface-100'], 0.07);
+  }
+
+  if (!('surface-300' in colors)) {
+    if ('surface-200' in themeColors) {
+      colors['surface-300'] = darkenColor(colors['surface-200'], 0.07);
+    } else {
+      colors['surface-300'] = darkenColor(colors['surface-100'], 0.14);
+    }
+  }
+
+  if (!('surface-content' in colors)) {
+    colors['surface-content'] = foregroundColor(colors['surface-100']);
+  }
+
+  const result = Object.fromEntries(
+    Object.entries(colors).map(([name, value]) => {
+      if (colorNames.includes(name)) {
+        // Add space separated color variables for each color
+        return [`--color-${name}`, colorVariableValue(value, colorSpace)];
+      } else {
+        // Additional properties such as `color-scheme` or CSS variable
+        return [name, value];
+      }
+    })
+  );
+
+  return result;
+}
+
+function round(value: number, decimals: number) {
+  if (value) {
+    return Number(value.toFixed(decimals));
+  } else {
+    return 0;
+  }
+}
+
+function isDark(color: Color | string) {
+  try {
+    if (wcagContrast(color, 'black') < wcagContrast(color, 'white')) {
+      return true;
+    }
+    return false;
+  } catch (e) {
+    return false;
+  }
+}
+
+/** Lighten or darken color based on contrast of input */
+function foregroundColor(color: Color | string, percentage = 0.8) {
+  try {
+    return isDark(color) ? lightenColor(color, percentage) : darkenColor(color, percentage);
+  } catch (e) {
+    // console.error('Unable to generate foreground color', color);
+  }
+}
+
+function lightenColor(color: Color | string, percentage: number) {
+  try {
+    return formatCss(interpolate([color, 'white'], 'oklch')(percentage));
+  } catch (e) {
+    // console.error('Unable to generate lighten color', color);
+  }
+}
+
+function darkenColor(color: Color | string, percentage: number) {
+  try {
+    return formatCss(interpolate([color, 'black'], 'oklch')(percentage));
+  } catch (e) {
+    // console.error('Unable to generate darken color', color);
+  }
+}
+
+/**
+ * Convert color to space separated components string
+ */
+function colorVariableValue(color: Color | string, colorSpace: 'rgb' | 'hsl' | 'oklch') {
+  try {
+    if (colorSpace === 'rgb') {
+      const { r, g, b } = rgb(color);
+      return `${round(r * 255, 6)} ${round(g * 255, 6)} ${round(b * 255, 6)}`;
+    } else if (colorSpace === 'hsl') {
+      const { h, s, l } = hsl(clampRgb(color));
+      return `${round(h, 6)} ${round(s * 100, 6)}% ${round(l * 100, 6)}%`;
+      // return `${round(h, 6)} ${round(Math.min(s * 100, 100), 6)}% ${round(Math.min(l * 100, 100), 6)}%`;
+    } else if (colorSpace === 'oklch') {
+      const { l, c, h } = oklch(color);
+      return `${round(l, 6)} ${round(c, 6)} ${round(h, 6)}`;
+    }
+  } catch (e) {
+    // console.error('Unable to convert color object to string', color);
+  }
+}
 
 /** Return a script tag that will set the initial theme from localStorage. This allows setting
  * the theme before anything starts rendering, even when SSR is in use.
